# docs/swagger/api-gateway/api_gateway_api_v1.yaml
openapi: 3.0.0
info:
  title: Wine API
  version: 1.0.0
paths:
  /api/v1/auth/login:
    post:
      summary: Log in
      requestBody:
        description: ID Token returned by Google
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                idToken:
                  type: string
                  example: "ID Token returned by Google"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT Token containing user data including Google ID
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '401':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid or expired token."
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "User not found"
...